+++ Changelog.md
@@ -42,6 +42,7 @@
 - Ensure consistent order when displaying test results (#6010)
 - Added Capybara Gem (#6012)
 - Separated peer review assignment upload/download from parent (#6048)
+- Expanded summary stats UI (#6050)
 - Fix bug for submission modal where students could submit files even if no file was chosen (#6052)
 
 ## [v2.0.10]